requirements.txt
pydantic
pyyaml
sphinx
sphinxcontrib-svg2pdfconverter
